+++ hedgehog/src/Yaya/Hedgehog/Expr.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE DeriveTraversable #-}
+{-# LANGUAGE StrictData #-}
 {-# LANGUAGE TemplateHaskell #-}
 
 module Yaya.Hedgehog.Expr where