+++ unsafe-test/test/Test/Fold.hs
@@ -2,10 +2,13 @@
 
 module Test.Fold where
 
+import           Data.Proxy
 import           Hedgehog
 import qualified Hedgehog.Gen as Gen
 
+import           Yaya.Fold
 import           Yaya.Fold.Common
+import           Yaya.Fold.Native
 import           Yaya.Hedgehog.Expr
 import           Yaya.Hedgehog.Fold
 import qualified Yaya.Unsafe.Fold.Instances ()
@@ -24,9 +27,11 @@ prop_fixCataRefl :: Property
 prop_fixCataRefl =
   property $ law_cataRefl =<< forAll (Gen.sized genFixExpr)
 
--- prop_fixCataCompose :: Property
--- prop_fixCataCompose =
---   property $ law_cataCompose size id =<< forAll (Gen.sized genFixExpr)
+prop_fixCataCompose :: Property
+prop_fixCataCompose =
+  property
+  $ law_cataCompose (Proxy :: Proxy (Fix Expr)) size id
+    =<< forAll (Gen.sized genFixExpr)
 
 -- | NB: Only in yaya-unsafe instead of yaya because the `Eq (Nu f)` instance is
 --       needed.
@@ -42,9 +47,11 @@ prop_nuCataRefl :: Property
 prop_nuCataRefl =
   property $ law_cataRefl =<< forAll (Gen.sized genNuExpr)
 
--- prop_nuCataCompose :: Property
--- prop_nuCataCompose =
---   property $ law_cataCompose size id =<< forAll (Gen.sized genNuExpr)
+prop_nuCataCompose :: Property
+prop_nuCataCompose =
+  property
+  $ law_cataCompose (Proxy :: Proxy (Nu Expr)) size id
+    =<< forAll (Gen.sized genNuExpr)
 
 tests :: IO Bool
 tests = checkParallel $$(discover)