+++ hedgehog/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to the [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
+## 0.1.2.0 – 2019–11–08
+### Added
+- `law_cataCompose`
+
 ## 0.1.1.0 – 2019–01–08
 ### Added
 - `law_anaRefl` as dual to `law_cataRefl`